requirements.txt
django>=4.0,<5.0
gunicorn>=20.0,<21.0
djangorestframework>=3
dj-database-url==1.0.0
psycopg2==2.9.5
python-dotenv==0.21.0
whitenoise>=6.2.0
pandas==1.5.1
requests==2.28.2